move 'slow' versions of logsumexp and mean_abs_diff to matutils
