+++ gensim/matutils.py
@@ -644,6 +644,52 @@ def qr_destroy(la):
     return q, r
 
 
+def logsumexp(x):
+    """
+    Log of sum of exponentials
+
+    Parameters
+    ----------
+    x : array_like
+        Input data
+
+    Returns
+    -------
+    float
+        log of sum of exponentials of elements in `x`
+
+    Notes
+    -----
+        for performance, does not support NaNs or > 1d arrays like
+        scipy.special.logsumexp()
+
+    """
+
+    x_max = np.max(x)
+    x = np.log(np.sum(np.exp(x - x_max)))
+    x += x_max
+
+    return x
+
+
+def mean_absolute_difference(a, b):
+    """
+    Mean absolute difference between two arrays
+
+    Parameters
+    ----------
+    a : (M,) array_like of float32
+    b : (M,) array_like of float32
+
+    Returns
+    -------
+    float32
+        mean(abs(a - b))
+    
+    """
+    return np.mean(np.abs(a - b))
+
+
 class MmWriter(object):
     """
     Store a corpus in Matrix Market format.
